feat(cli): add --unique-pact-file-names option
This allows multiple mock services to be run in parallel for the same consumer/provider
pair on different ports without the pact files overwriting each other.
Note that the developer must ensure that the pact-dir is emptied before each
test run to ensure that the files do not accumulate between runs.